refactor:컴포넌트 파일 정리(_styles 폴더랑 _components 폴더 분 리) (#2)
