+++ src/app/company/[id]/manage_camp/add_camp/_components/AddressModal.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React from 'react';
 import DaumPostcodeEmbed, { Address } from 'react-daum-postcode';
-import styles from './searchAddress.module.css';
+import styles from '../_styles/SearchAddress.module.css';
 
 interface Props {
   handleCompleteAddress: (data: Address) => void;
